Update FutureStuff.md
Added TYPES of windows, reliability for Food for Thought, this is my take on how we solve a major problem we have: users not organizing their tabs underneath the respected windows. Using types and notification to notify them that they should be doing this in the right places. 

The question to consider: How to make sure that users open different TYPES of windows and stick to those windows category???