All user browisng session is stored in a revealing module pattern. This allows for easier manipulation with the currently opened windows and tabs. Need to create Launch and Wormhole tabs for each window...
Currently all tabs are stored regardless of whether the windows are linked to a book or not. So when users link to a book, the wormhole and launch will be automatically populated according to their window's browsing session. Why do this? In the FUTURE, we can distinguish between links searched during a window being linked to a book, vs a linked search when window is unlinked to any book.